Implement feature add and edit
Tweak add and edit features from AB3 to be more aligned with InternSHIP add and edit feature. This is a rough implementation with several areas of improvement.

For example, deadline field is currently implemented as a string and there is no error checking for date-formatting. For add feature, the tag with default "Pending" will not be added automatically.
The corresponding test needs further updating as well.

Refactor codes to make subsequent implementation easier.